PBP 91426 ZSF CallLocation refactor (#19301)
