+++ spec/lib/zero_silent_failures/monitor_spec.rb
@@ -1,12 +1,13 @@
 # frozen_string_literal: true
 
 require 'rails_helper'
+require 'logging/call_location'
 require 'zero_silent_failures/monitor'
 
 RSpec.describe ZeroSilentFailures::Monitor do
   let(:service) { 'test-application' }
   let(:monitor) { described_class.new(service) }
-  let(:call_location) { described_class::CallLocation.new('fake_func', 'fake_file', 'fake_line_42') }
+  let(:call_location) { Logging::CallLocation.new('fake_func', 'fake_file', 'fake_line_42') }
   let(:tags) { ["service:#{service}", "function:#{call_location.base_label}"] }
   let(:user_account_uuid) { '123-test-uuid' }
   let(:additional_context) { { test: 'foobar' } }
@@ -22,19 +23,12 @@ RSpec.describe ZeroSilentFailures::Monitor do
     }
   end
 
-  describe '::CallLocation' do
-    it 'responds to and returns expected values' do
-      expect(call_location.base_label).to eq('fake_func')
-      expect(call_location.path).to eq('fake_file')
-      expect(call_location.lineno).to eq('fake_line_42')
-    end
-  end
-
   context 'with a call location provided' do
     describe '#log_silent_failure' do
       it 'logs a silent failure with call location' do
         payload[:statsd] = 'silent_failure'
 
+        expect(monitor).to receive(:parse_caller).with(call_location).and_call_original
         expect(StatsD).to receive(:increment).with('silent_failure', tags:)
         expect(Rails.logger).to receive(:error).with('Silent failure!', payload)
 
@@ -46,6 +40,7 @@ RSpec.describe ZeroSilentFailures::Monitor do
       it 'logs a silent failure with call location and no confirmation' do
         payload[:statsd] = 'silent_failure_avoided_no_confirmation'
 
+        expect(monitor).to receive(:parse_caller).with(call_location).and_call_original
         expect(StatsD).to receive(:increment).with('silent_failure_avoided_no_confirmation', tags:)
         expect(Rails.logger).to receive(:error).with('Silent failure avoided (no confirmation)', payload)
