+++ spec/sidekiq/central_mail/submit_form4142_job_spec.rb
@@ -173,7 +173,7 @@ RSpec.describe CentralMail::SubmitForm4142Job, type: :job do
                 form526_submission_id: form526_submission.id
               },
               nil,
-              call_location: instance_of(ZeroSilentFailures::Monitor::CallLocation)
+              call_location: instance_of(Logging::CallLocation)
             )
 
             args = { 'jid' => form526_job_status.job_id, 'args' => [form526_submission.id] }
